Consolidate FB launch configurations
