--- plugins/org.eclipse.fordiac.ide.debug.ui/src/org/eclipse/fordiac/ide/debug/ui/simple/SimpleFBLaunchConfigurationTabGroup.java
@@ -1,26 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2022 Martin Erich Jobst
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *   Martin Jobst - initial API and implementation and/or initial documentation
- *******************************************************************************/
-package org.eclipse.fordiac.ide.debug.ui.simple;
-
-import org.eclipse.debug.ui.AbstractLaunchConfigurationTabGroup;
-import org.eclipse.debug.ui.CommonTab;
-import org.eclipse.debug.ui.ILaunchConfigurationDialog;
-
-public class SimpleFBLaunchConfigurationTabGroup extends AbstractLaunchConfigurationTabGroup {
-
-	@Override
-	public void createTabs(final ILaunchConfigurationDialog dialog, final String mode) {
-		setTabs(new SimpleFBLaunchConfigurationTab(), new CommonTab());
-	}
-
-}
