--- plugins/org.eclipse.fordiac.ide.debug/src/org/eclipse/fordiac/ide/debug/simple/SimpleFBLaunchConfigurationAttributes.java
@@ -1,19 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2022 Martin Erich Jobst
- * 
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- * 
- * SPDX-License-Identifier: EPL-2.0
- * 
- * Contributors:
- *   Martin Jobst - initial API and implementation and/or initial documentation
- *******************************************************************************/
-package org.eclipse.fordiac.ide.debug.simple;
-
-import org.eclipse.fordiac.ide.debug.fb.FBLaunchConfigurationAttributes;
-
-public interface SimpleFBLaunchConfigurationAttributes extends FBLaunchConfigurationAttributes {
-	String ID = "org.eclipse.fordiac.ide.debug.simpleFBLaunch";
-}
